requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
pushbullet.py==0.12.0